Remove sender field when unstaking/claiming
The SenderFormField is now wrapped in a conditional check (mode === TransactionMode.DELEGATE), ensuring it is only rendered when the user is performing a delegation.

For both unstaking and claiming, the user only needs to select the StakingPosition, and the sender is implicitly part of that position. Thus, the form becomes more streamlined and easier to use for these transaction modes.